chore: remove some log and move function to custom service
