+++ src/api/purchase/services/extended.js
@@ -8,7 +8,6 @@ const {
   userExtendedService,
   purchaseService,
   userService,
-  couponExtendedService,
   couponService,
 } = require("../../../utils/services");
 
@@ -22,8 +21,6 @@ module.exports = ({ strapi }) => ({
 
     const { user, subscription, coupon } = purchase;
     const now = new Date();
-
-    console.log(coupon);
 
     let subscribeTime = user.subscribedUntil
       ? new Date(user.subscribedUntil)
@@ -55,34 +52,6 @@ module.exports = ({ strapi }) => ({
         data: { used: coupon.used + 1 },
       });
     }
-
-    return purchase;
-  },
-
-  async update(purchaseId) {
-    let purchase = await purchaseService().findOne(purchaseId, {
-      populate: "*",
-    });
-
-    const { coupon, user, subscription } = purchase;
-
-    if (await couponExtendedService().isValid(coupon, user)) {
-      const price = couponExtendedService().offAmount(subscription, coupon);
-
-      if (price !== purchase.price) {
-        purchase = await purchaseService().update(purchase.id, {
-          data: { price },
-          populate: "*",
-        });
-      }
-
-      return purchase;
-    }
-
-    purchase = await purchaseService().update(purchase.id, {
-      data: { coupon: null, price: subscription.currentPrice },
-      populate: "*",
-    });
 
     return purchase;
   },
